Cleanup: move requestQueue to separate file, rename variables
